+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/):
 
 ### Added
 
+- Tree view for the preset archive previewer ([#3525])
 - Support compressed tarballs (`.tar.gz`, `.tar.bz2`, etc.) in the preset archive previewer ([#3518])
 - New `Path.os()` API creates an OS-native `Path` ([#3541])
 
@@ -1596,6 +1597,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/):
 [#3494]: https://github.com/sxyazi/yazi/pull/3494
 [#3514]: https://github.com/sxyazi/yazi/pull/3514
 [#3518]: https://github.com/sxyazi/yazi/pull/3518
+[#3525]: https://github.com/sxyazi/yazi/pull/3525
 [#3532]: https://github.com/sxyazi/yazi/pull/3532
 [#3540]: https://github.com/sxyazi/yazi/pull/3540
 [#3541]: https://github.com/sxyazi/yazi/pull/3541